requirements.txt
Flask==3.1.1
joblib==1.2.0
numpy==1.23.5
pandas==1.5.2
python-dotenv==1.1.0
Requests==2.32.3
scikit_learn==1.1.3
spotipy==2.25.1
